LayoutTests/platform/chromium/fast/events/intercept-postmessage.html

<html>
  <head>
    <script>
      function write(str) {
        pre = document.getElementById('console');
        text = document.createTextNode(str + '\n');
        pre.appendChild(text);
      }

      function recvMsg1(e) {
        write("RECV1");
        window.removeEventListener("message", recvMsg1, false);

        if (window.layoutTestController)
          layoutTestController.interceptPostMessage = true;

        window.addEventListener("message", recvMsg2, false);
        window.postMessage("Message 2", "*");
      
        // Ensure that we're intercepting postMessages before the origin check
        window.postMessage("Message 3", "http://example.org"); 

        // We need to call setTimeout here because we intercept the next event.
        // If this test becomes flaky, you may need to increase this timeout.
        window.setTimeout("done()", 50);
      }

      function recvMsg2(e) {
        write("RECV2");
      }

      function done() {
        if (window.layoutTestController)
          window.layoutTestController.notifyDone();
      }

      window.addEventListener("message", recvMsg1, false);

      if (window.layoutTestController) {
        window.layoutTestController.dumpFrameLoadCallbacks();
        window.layoutTestController.dumpAsText();
        window.layoutTestController.waitUntilDone();
      }

      window.postMessage("Message 1", "*");
    </script>
  </head>
  <body><pre id="console"></pre></body>
</html>
